requirements.txt
asgiref==3.3.3
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djoser
djangorestframework-simplejwt==4.8.0
Pillow==9.5.0